refactor: divide socket handlers
add socket handler directory in services
