+++ server/src/services/socketHandlers/index.js
@@ -1,0 +1,8 @@
+const lobbyHandler = require("./lobby");
+const reactionHandler = require("./reaction");
+module.exports = (io) => {
+    io.on("connection", (socket) => {
+    lobbyHandler(socket);
+    reactionHandler(socket);
+    });
+}
