+++ server/src/services/socketHandlers/reaction.js
@@ -1,0 +1,1 @@
+module.exports = (socket) => {}
